requirements.txt
discord.py
youtube-dl
python-dotenv~=0.19.0
